+++ taf/api/roles.py
@@ -655,7 +655,7 @@ def list_keys_of_role(
     auth_repo = AuthenticationRepository(path=path)
     key_ids = auth_repo.get_role_keys(role=role)
     for key_id in key_ids:
-        print(get_metadata_key_info(auth_repo.certs_dir, key_id))
+        print(f"\n{get_metadata_key_info(auth_repo.certs_dir, key_id)}")
 
 
 @log_on_start(DEBUG, "Removing role {role:s}", logger=taf_logger)