+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning][semver].
 
 ### Added
 
+- List keys of roles with additional information read from certificates command ([355])
+- Export certificate from the inserted YubiKey ([355])
 - Add signing keys given a public key when creating a new authentication repository ([354])
 - Allow specification of names of YubiKeys in repository description json ([354])
 - Model repository description json input using `attrs` and `cattrs` and its validation ([354])
@@ -35,6 +37,7 @@ and this project adheres to [Semantic Versioning][semver].
 - Fix commits per repositories function when same target commits are on different branches ([337])
 - Add missing `write` flag to `taf targets sign` ([329])
 
+[355]: https://github.com/openlawlibrary/taf/pull/355
 [354]: https://github.com/openlawlibrary/taf/pull/354
 [353]: https://github.com/openlawlibrary/taf/pull/353
 [352]: https://github.com/openlawlibrary/taf/pull/352